+++ R/reset.R
@@ -9,15 +9,21 @@
 #'
 #' @return Invisibly returns the vector of deleted files.
 #' @export
-reset <- function() {
+reset <- function(interactive = NULL) {
   files <- get_all_metadata_files()
+
+  if (is.null(interactive)) {
+    interactive = base::interactive()
+  } else {
+    interactive = match(interactive, c(TRUE,FALSE))
+  }
 
   if (length(files) == 0) {
     message("No files found.")
     return(invisible(character()))
   }
 
-  if (interactive()) {
+  if (isTRUE(interactive)) {
     cat("Files to be deleted:\n")
     cat(paste0(" - ", files), sep = "\n")
     choice <- utils::menu(c("Yes", "No"), title = "Delete all files above?")