+++ tests/testthat/test-reset.R
@@ -4,8 +4,8 @@ test_that("reset deletes reproducibility files", {
   oldwd <- setwd(tmp)
   on.exit(setwd(oldwd), add = TRUE)
 
-  prefix <- reproducibleRchunks::default_prefix()
-  filetype <- reproducibleRchunks::default_filetype()
+  prefix <- default_prefix()
+  filetype <- default_filetype()
 
   f1 <- file.path(tmp, paste0(prefix, "_test1.", filetype))
   f2 <- file.path(tmp, paste0(prefix, "_test2.", filetype))
@@ -14,7 +14,7 @@ test_that("reset deletes reproducibility files", {
 
   expect_true(all(file.exists(c(f1, f2))))
 
-  reproducibleRchunks::reset(interactive = FALSE)
+  reproducibleRchunks::reset()
 
   expect_false(any(file.exists(c(f1, f2))))
 })